BE:fix: update MemberLevelInfo id type from Long to Integer - database column is TINYINT

+++ backend/src/main/java/org/lightnsalt/hikingdom/domain/member/repository/MemberLevelInfoRepository.java
@@ -3,6 +3,6 @@ package org.lightnsalt.hikingdom.domain.member.repository;
 import org.lightnsalt.hikingdom.domain.member.entity.MemberLevelInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberLevelInfoRepository extends JpaRepository<MemberLevelInfo, Long> {
+public interface MemberLevelInfoRepository extends JpaRepository<MemberLevelInfo, Integer> {
 
 }
